# server/openapi.yaml
openapi: 3.0.3
info:
  title: Huuk API (by teenueng)
  version: 0.1.0
  description: >-
    REST API for Huuk by teenueng (Vue 3 frontend in `app/`,
    Rust/Axum backend in `server/`). All request and response bodies are JSON
    with camelCase field names; errors use the envelope
    `{ "error": "message" }`.

    Authentication: send `Authorization: Bearer <token>` from
    `POST /api/auth/login`. Auth is only enforced when `authRequired` is true
    (production default); in demo mode (`DEMO_MODE=1`) reads and writes are
    open and the Bearer token is optional. In production every `GET` also
    requires a live token, except liveness (`/api/health`, `/api/ready`), the
    redacted `GET /api/setup`, the public delivery API (`/api/public/*`), the
    brand CI (`/api/brand`, `/api/brand/fonts`, font files) and the OAuth/docs
    routes; every mutating operation is gated by the permission named in its
    `x-permission` field.

    Content items (`/api/content`) implement an editorial workflow with
    optimistic concurrency: send the `version` you last read on every
    update/publish/schedule/restore call and expect `409` when someone else
    saved in between.

    Interactive docs are served by the app itself: `GET /api/docs` (Swagger UI)
    and `GET /api/openapi.yaml` (this document).
servers:
  - url: /
    description: Same origin (the server also serves the built SPA when STATIC_DIR is set)
  - url: http://localhost:8787
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/setup:
    get:
      operationId: getSetup
      tags: [Setup]
      summary: Get the workspace configuration
      security: []
      responses:
        '200':
          description: Current SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
    patch:
      operationId: patchSetup
      tags: [Setup]
      summary: Update workspace configuration
      description: >-
        Partial update. A rejected `roles` replacement never half-applies.
        Replacing roles must be non-empty, duplicate-free, permission-valid and
        keep every role a directory user holds.
      x-permission: setup.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupPatch' }
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/setup/options:
    post:
      operationId: addOption
      tags: [Setup]
      summary: Add an item to an option list
      description: Duplicates are ignored.
      x-permission: setup.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddOptionBody' }
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/setup/users:
    post:
      operationId: addUser
      tags: [Setup]
      summary: Add a directory user
      x-permission: users.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserBody' }
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/setup/users/{name}:
    delete:
      operationId: removeUser
      tags: [Setup]
      summary: Remove a directory user
      x-permission: users.manage
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/setup/roles:
    post:
      operationId: addRole
      tags: [Setup]
      summary: Add a role
      x-permission: users.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleBody' }
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/setup/roles/{name}:
    delete:
      operationId: removeRole
      tags: [Setup]
      summary: Remove a role
      description: The last role and roles assigned to directory users cannot be removed.
      x-permission: users.manage
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200':
          description: Updated SetupConfig.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupConfig' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Register an account (auto-login)
      description: >-
        Public endpoint. Requires `allowRegistration`; the new account gets the
        directory role when the name exists, otherwise Viewer.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterBody' }
      responses:
        '200':
          description: Session token and user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403':
          description: Registration is disabled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already registered or display name already taken.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Log in with email and password
      description: >-
        On success returns a 64-char session token (30-day TTL). Repeated
        failures lock the account for 15 minutes (`429`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginBody' }
      responses:
        '200':
          description: Session token and user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid email or password (message is identical for both).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/auth/me:
    get:
      operationId: me
      tags: [Auth]
      summary: Get the current session's user
      responses:
        '200':
          description: The signed-in user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/AuthUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Revoke the current session
      responses:
        '200':
          description: Session revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout-all:
    post:
      operationId: logoutAll
      tags: [Auth]
      summary: Revoke every session of the account
      responses:
        '200':
          description: All sessions revoked, including the caller's.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [Auth]
      summary: Change the current account's password
      description: Keeps the caller's token valid and revokes every other session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordBody' }
      responses:
        '200':
          description: Password changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/profile:
    patch:
      operationId: updateProfile
      tags: [Auth]
      summary: Rename the current account
      description: >-
        Requires a live session but no permission. The directory entry follows
        the rename so the role survives; active post locks are rewritten.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileBody' }
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/AuthUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/auth/accounts:
    get:
      operationId: listAccounts
      tags: [Accounts]
      summary: List accounts with live session counts
      description: Never exposes password hashes.
      x-permission: users.manage
      responses:
        '200':
          description: Account rows sorted by name.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AccountRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAccount
      tags: [Accounts]
      summary: Onboard an account (admin-created)
      description: >-
        Creates the sign-in account and assigns the directory role. When
        `password` is omitted the server generates one and returns it exactly
        once as `temporaryPassword`.
      x-permission: users.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewAccountBody' }
      responses:
        '200':
          description: Account created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedAccount' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/auth/accounts/{email}:
    delete:
      operationId: deleteAccount
      tags: [Accounts]
      summary: Delete an account and revoke its sessions
      description: The directory entry is untouched; the last administrator cannot be deleted.
      x-permission: users.manage
      parameters:
        - name: email
          in: path
          required: true
          description: Account email (case-insensitive).
          schema: { type: string, format: email }
      responses:
        '200':
          description: Account deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/content:
    get:
      operationId: listContent
      tags: [Content]
      summary: List content items (summaries)
      description: Sorted by `updatedAt` descending; due scheduled items are promoted first.
      security: []
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by workflow status.
          schema:
            type: string
            enum: [draft, review, scheduled, published, archived]
        - name: q
          in: query
          required: false
          description: Case-insensitive search over title, excerpt and tags.
          schema: { type: string }
      responses:
        '200':
          description: Content summaries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentSummary' }
    post:
      operationId: createContent
      tags: [Content]
      summary: Create a content item
      description: >-
        Server assigns `id`, `slug` (when empty), timestamps and version; the
        status is forced to `draft` unless it is `draft` or `review`.
      x-permission: content.write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentInput' }
      responses:
        '200':
          description: Created content item (revision bodies omitted).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/content/{id}:
    get:
      operationId: getContent
      tags: [Content]
      summary: Get a content item
      security: []
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content item (revision bodies omitted).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateContent
      tags: [Content]
      summary: Update a content item
      description: >-
        Requires the current `version`; the pre-edit state is stored as a
        revision (max 20). `status` may only be set to draft, review or
        archived here — publishing and scheduling have their own endpoints.
      x-permission: content.write
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentPatch' }
      responses:
        '200':
          description: Updated content item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Version conflict (content changed since the version you sent).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      operationId: deleteContent
      tags: [Content]
      summary: Delete a content item
      x-permission: content.delete
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/content/{id}/publish:
    post:
      operationId: publishContent
      tags: [Content]
      summary: Publish a content item
      x-permission: content.publish
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionBody' }
      responses:
        '200':
          description: Published content item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/content/{id}/unpublish:
    post:
      operationId: unpublishContent
      tags: [Content]
      summary: Unpublish a content item (back to draft)
      x-permission: content.publish
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionBody' }
      responses:
        '200':
          description: Unpublished content item (status draft).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token from `POST /api/auth/login`. Optional while
        `authRequired` is false (demo mode); required for mutations in
        production.
...